Drop tiny.json and use "small" library option instead.
This lead to same compilation result if "release" profile is used.
When "target.default_lib": "small" option is used in mbed_app.json,
that leads to GCC to add --spec=nano.spec option which uses
newlib-nano instead of normal newlib.
Our tiny.json was doing this, but now Mbed OS supports that option
so we don't need to have separate build profile.