configs:no_rot: Add configration for MIMXRT1050_EVK
Add configration for MIMXRT1050_EVK in internal_flash_no_rot.json.
With the Soc security modules like HAB, BEE, DCP and others,
the physical external flash can be treated as internal flash
with the same security feature.

The configration supports MIMXRT1050_EVK 8MB QSPI flash in default.
With the FLASHIAP and the default flash memory map, it can support
kvstore and OTA features.

For the Hyper flash, please delete the line:
	"target.macros_remove" : ["HYPERFLASH_BOOT"]
and adjust the other settings to match the Hyper flash size and
flash map.

Signed-off-by: Gavin Liu <gang.liu@nxp.com>